requirements.txt
langchain-community
langchain_experimental
langchain_google_genai
streamlit
arxiv